Update the jupyter container
This commit updates many of the packages in our Jupyter notebook image.
Some of them were getting pretty old.  Unfortunately, the version of
Proj packaged in debian is very very old, so updating pyproj (which was
two years old) requires compiling Proj from source.  This makes the
process veeeery long.

If we were to switch everything to conda this would become much simpler.