Differentiate log configuration of the web app during testing
A number of our loggers are getting disabled during testing.  In an
(unsuccessful) effort to avoid this, this commit changes the tdmq app
logic to only apply its logging configuration if it isn't in testing
mode (i.e., config['TESTING'] is not True).

Also, to bluntly but successfully work around the issue, the commit
introduces a new fixture that enables all loggers before all tests.